+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 [0.0.9] - 2023-01-18
 --------------------
+* Deprecations
+  * Added warnings for the F10.7 and Kp tags that load data belonging in
+  their own Instruments
 * Enhancements
    * Added an instrument for the LASP MgII core-to-wing index
    * Added functions for general LISIRD downloads